Outline the new IODag type.
This type allows for edges that come/go to/from infinity (null),
instead of just between nodes. This solves a common problem where DAGs
need to have a 'root' and the edges to that root are subject to
different behaviors (weights).

Additionally, the NodeHandles and internal data structures are intended
to be serializable.

Edges are never gc'd, but Nodes with no attached edges *are* (or
rather, will be).